backend/requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.1.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
datasets==3.3.2
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
faiss-cpu==1.10.0
farasapy==0.0.14
fastapi==0.115.8
fastapi-csrf-protect==1.0.2
fastapi-mail==1.4.2
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langchain-together==0.3.0
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
motor==3.7.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==1.26.4
openai==1.65.4
openpyxl==3.1.2
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pillow==11.1.0
polars==1.23.0
propcache==0.3.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
striprtf==0.0.26
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tnkeeh==0.0.9
together==1.4.1
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
upstash-redis==1.3.0
gradio_client==1.8.0
passlib==1.7.4